Add compile order option to the bloop config
The compile order is important because many projects define their own
compile order and that changes the semantics of compilation.

Let's add this option under a `CompileOptions` field in our bloop config
file. This new field will in the future, before 1.0.0, host
`classpathOptions` which seems kinda out of place. But for now, we
abstain ourselves from making this breaking change.