Add `nativeConfig` to `Config`
This replaces the previous `nativeClasspath`. `nativeConfig` contains
all the configuration that should be required for a Scala Native
project, including the classpath of the Scala Native toolchain.

This configuration item is currently optional. If it is not specified,
then the Scala Native toolchain will be resolved using Coursier and the
default values will be used for configuring the toolchain.

At the moment, our sbt and Maven plugin never populate this
configuration.