Use circe instead of metaconfig
The use of metaconfig slows down the startup considerably: it takes 60
seconds to load the beefy json configuration files from the lila
project.

As speed the speed of project load is critical, we cannot afford to use
metaconfig for this task, so we switch directly to circe. With circe,
loading lichess is instant (less than one second)!