Redesign how platform is modeled in the config
Instead of having one field per platform, we use the fact that a target
can only have one platform to create a field `platform` that contains
all the information about that platform.

That helps us make the design more elegant and remove the somewhat
ad-hoc `jvm` field that we had before.

Aside from this, we now check that our dear Scala 2.10 variant of the
encoders/decoders work. Note that we cannot avoid the duplication with
`ConfigEncoderDecoders` because 2.12 uses `circe-derivation` (which
doesn't support Scala 2.10) and 2.10 uses circe semiauto.