Support runtime classpath for JVM projects
Most build tools use a different classpath for compiling and running JVM
projects. This commit adds fields to represent the runtime classpath of
JVM projects in the Bloop configuration file.

The following fields are added to the JVM platform:

 * `classpath`, which represents the runtime classpath of the project.
 * `resources`, which represents the resources available to the project
   at runtime.

When one of these fields is missing, Bloop will fallback to the field
with the same name in project definition.

Fixes #1233