deps: bump maven related deps
I'm going to manually test this on a few repos locally, but this fixes
the issue with the Maven api not being marked as provided. Once you bump
the plugin-plugin deps it actually yells pretty loud at you if you don't
do this, so I wanted to ensure we could. This ensures that Maven is on
the latest and also the maven plugin plugin tools.