fix: ensure we java modules are still working
There was a change I didn't catch in the previous pr due to how I was
testing. I wasn't being explicit about the version when running
`bloopInstall` so it was using the last published version. This fixes
that by using the version that is inside of the pom when you test.

The other issue, which the title is about, is that
`mojo.getScalacArgs()` can throw now since it attempts to get the scala
version set for the module. However, in Java modules there isn't one, so
this was blowing up. This fixes that by wrapping that call like we
earlier for other calls that can throw.

This also fixes some pom settings that were still pointing to the old
repo.